+++ src/Rest/Routes/Integrations/Greenhouse/FormSubmitGreenhouseRoute.php
@@ -81,15 +81,8 @@ class FormSubmitGreenhouseRoute extends AbstractFormSubmit
 	 */
 	protected function submitAction(array $formDetails)
 	{
-		$formId = $formDetails[UtilsConfig::FD_FORM_ID];
-
 		// Send application to Greenhouse.
-		$response = $this->greenhouseClient->postApplication(
-			$formDetails[UtilsConfig::FD_ITEM_ID],
-			$formDetails[UtilsConfig::FD_PARAMS],
-			$formDetails[UtilsConfig::FD_FILES],
-			$formId
-		);
+		$response = $this->greenhouseClient->postApplication($formDetails);
 
 		$formDetails[UtilsConfig::FD_RESPONSE_OUTPUT_DATA] = $response;