+++ src/Rest/Routes/Integrations/Pipedrive/FormSubmitPipedriveRoute.php
@@ -81,14 +81,8 @@ class FormSubmitPipedriveRoute extends AbstractFormSubmit
 	 */
 	protected function submitAction(array $formDetails)
 	{
-		$formId = $formDetails[UtilsConfig::FD_FORM_ID];
-
 		// Send application to Hubspot.
-		$response = $this->pipedriveClient->postApplication(
-			$formDetails[UtilsConfig::FD_PARAMS],
-			$formDetails[UtilsConfig::FD_FILES],
-			$formId
-		);
+		$response = $this->pipedriveClient->postApplication($formDetails);
 
 		$formDetails[UtilsConfig::FD_RESPONSE_OUTPUT_DATA] = $response;